refactor of blocksparse
