+++ tensornetwork/backends/symmetric/decompositions.py
@@ -13,8 +13,10 @@
 # limitations under the License.
 """Tensor Decomposition Implementations."""
 from typing import Optional, Any, Tuple
-from tensornetwork.block_sparse.utils import (
-    _find_transposed_diagonal_sparse_blocks, get_real_dtype, SIZE_T)
+from tensornetwork.block_sparse.blocksparse_utils import (
+    _find_transposed_diagonal_sparse_blocks, SIZE_T)
+from tensornetwork.block_sparse.utils import get_real_dtype
+
 from tensornetwork.block_sparse.blocksparsetensor import (BlockSparseTensor,
                                                           ChargeArray)
 import numpy as np
