+++ tensornetwork/block_sparse/charge.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import numpy as np
+from tensornetwork.block_sparse.utils import intersect
 from typing import (List, Optional, Type, Any, Union, Callable)
 
 #TODO (mganahl): switch from column to row order for unique labels
@@ -676,72 +677,6 @@ def fuse_ndarray_charges(charges_A: np.ndarray, charges_B: np.ndarray,
   return np.stack(comb_charges, axis=1)
 
 
-def intersect(A: np.ndarray,
-              B: np.ndarray,
-              axis=0,
-              assume_unique=False,
-              return_indices=False) -> Any:
-  """
-  Extends numpy's intersect1d to find the row or column-wise intersection of
-  two 2d arrays. Takes identical input to numpy intersect1d.
-  Args:
-    A, B (np.ndarray): arrays of matching widths and datatypes
-  Returns:
-    ndarray: sorted 1D array of common rows/cols between the input arrays
-    ndarray: the indices of the first occurrences of the common values in A.
-      Only provided if return_indices is True.
-    ndarray: the indices of the first occurrences of the common values in B.
-      Only provided if return_indices is True.
-  """
-  # pylint: disable=line-too-long
-  # see
-  # https://stackoverflow.com/questions/8317022/ get-intersecting-rows-across-two-2d-numpy-arrays
-  #pylint: disable=no-else-return
-  A = np.ascontiguousarray(A)
-  B = np.ascontiguousarray(B)
-  if A.ndim != B.ndim:
-    raise ValueError("array ndims must match to intersect")
-  if A.ndim == 1:
-    return np.intersect1d(
-        A, B, assume_unique=assume_unique, return_indices=return_indices)
-
-  elif A.ndim == 2:
-    if axis == 0:
-      ncols = A.shape[1]
-      if A.shape[1] != B.shape[1]:
-        raise ValueError("array widths must match to intersect")
-
-      dtype = {
-          'names': ['f{}'.format(i) for i in range(ncols)],
-          'formats': ncols * [A.dtype]
-      }
-      if return_indices:
-        C, A_locs, B_locs = np.intersect1d(
-            A.view(dtype),
-            B.view(dtype),
-            assume_unique=assume_unique,
-            return_indices=return_indices)
-        return C.view(A.dtype).reshape(-1, ncols), A_locs, B_locs
-      C = np.intersect1d(
-          A.view(dtype), B.view(dtype), assume_unique=assume_unique)
-      return C.view(A.dtype).reshape(-1, ncols)
-
-    elif axis == 1:
-      out = intersect(
-          A.T.copy(),
-          B.T.copy(),
-          axis=0,
-          assume_unique=assume_unique,
-          return_indices=return_indices)
-      if return_indices:
-        return out[0].T, out[1], out[2]
-      return out.T
-
-    raise NotImplementedError(
-        "intersection can only be performed on first or second axis")
-
-  raise NotImplementedError("intersect is only implemented for 1d or 2d arrays")
-
 
 def fuse_charges(charges: List[BaseCharge], flows: List[bool]) -> BaseCharge:
   """
@@ -763,44 +698,6 @@ def fuse_charges(charges: List[BaseCharge], flows: List[bool]) -> BaseCharge:
   for n in range(1, len(charges)):
     fused_charges = fused_charges + charges[n] * flows[n]
   return fused_charges
-
-
-def fuse_degeneracies(degen1: Union[List, np.ndarray],
-                      degen2: Union[List, np.ndarray]) -> np.ndarray:
-  """
-  Fuse degeneracies `degen1` and `degen2` of two leg-charges
-  by simple kronecker product. `degen1` and `degen2` typically belong to two
-  consecutive legs of `BlockSparseTensor`.
-  Given `degen1 = [1, 2, 3]` and `degen2 = [10, 100]`, this returns
-  `[10, 100, 20, 200, 30, 300]`.
-  When using row-major ordering of indices in `BlockSparseTensor`,
-  the position of `degen1` should be "to the left" of the position of `degen2`.
-  Args:
-    degen1: Iterable of integers
-    degen2: Iterable of integers
-  Returns:
-    np.ndarray: The result of fusing `dege1` with `degen2`.
-  """
-  return np.reshape(
-      np.array(degen1)[:, None] * np.array(degen2)[None, :],
-      len(degen1) * len(degen2))
-
-
-def fuse_ndarrays(arrays: List[Union[List, np.ndarray]]) -> np.ndarray:
-  """
-  Fuse all `arrays` by simple kronecker addition.
-  Arrays are fused from "right to left", 
-  Args:
-    arrays: A list of arrays to be fused.
-  Returns:
-    np.ndarray: The result of fusing `arrays`.
-  """
-  if len(arrays) == 1:
-    return np.array(arrays[0])
-  fused_arrays = np.asarray(arrays[0])
-  for n in range(1, len(arrays)):
-    fused_arrays = np.ravel(np.add.outer(fused_arrays, arrays[n]))
-  return fused_arrays
 
 
 def charge_equal(c1: BaseCharge, c2: BaseCharge) -> bool:
